test(window): retain only test_basic.py for window tests
